Correct default policy_file in the sample config
When the sample config is generated, the policy_file default value
specified by oslo.policy is showing up, not the default value that
Cinder is using.  Correct this by enhancing the function specified
for cinder in the oslo.config.opts.defaults entry point.

This patch exposed Bug #1807184 in oslo.policy.  The requirements
and lower-constraints changes on this patch bump the minimum
version of oslo.policy we use to include the fix (1.44.1).  This
version was already included in upper-constraints by Change
I50bb9d48106376f5baaafae372492c0744182da2.

Change-Id: Ic28d660a9d876139f47f98cf1024a30b005cd0e0
Closes-bug: #1807071