Revert "Ceph driver should respect the `--incremental` option for backups"
This reverts commit 797c5aa78574a29a9a0bf7dd1b8adb47554b42c2.

That patch doesn't really fix the Ceph full backup issue.

When doing a non incremental Ceph to Ceph backup, instead of doing a
full backup that is ussable by the incremental feature, it creates a
full backup as if it where a backup from a non Ceph volume, which then
cannot be used by the incremental feature.

Change-Id: I73a40babf5cb13956e44d4caef3d0ff7ef698048