Move 'attach.end' notify to attachment_complete
Since microversion 3.44, we update `attachemnt_update` to return
connector and make volume to 'attaching', and add a new api
`attachment_complete` to change volume status to 'in-use', that means
the end of attachment is changed to `attachment_complete`.

But the 'attach.end' notify is still send in attachment_update, which
is unexpected. So, in this patch we move 'attach.end' notify to
`attachment_complete`.

Change-Id: I86dad7afb52d2d253799b543f484e2d55f6e9e6d
Closes-bug: #1808941