VMAX driver doc - notification of fix on AFA OS upgrade issue
Workload support was dropped in PowerMax OS 5978. A fix has
been merged to Rocky to support a PowerMax OS 5878 upgrade
on a VMAX All Flash, so all functionality on existing OpenStack
Instances and Volumes created from VMAX All Flash storage, will
work as before. This is not an issue in the (not yet released)
Stein as the fix is in already in place, so does not need to be
documented.

Change-Id: I698367020d5f10900bb4f6347f21154254ee5bee