Handle 'is_public' filter in volume-type list
If the resource is `volume_type` then `is_public=None`
is the default filter passed with it.
This filter should be handled by the general filtering
function which was included to volume type list code
introduced in MV 3.52.
If there is only one invalid_filter (is_public) and
the resource is volume_type then no exception should be
raised.

Change-Id: I7759a24895d5dac4306263712a40711cb8d06c7c
Closes-Bug: #1808620