Fix CI failed: test_get_volume_filter_options_using_config
Recently, CI failed(py27, py35 and other unit tests) due to
test_get_volume_filter_options_using_config case.

In patch[1], we remove deprecated query_volume_filters config option,
but the test case is fixed by a wrong way. The
``CONF.resource_query_filters_file`` is a **json path**, but NOT a
**json str**, if we want to overwrite it, we should overwrite the
_FILTERS_COLLECTION.[2]

Unfortunately, if we have already called the _initialize_filters in
other case (before this case), this error will not be exposed. So, we
didn't find this error occurring in first-CI.

This patch try to fix this issue.

Co-Authored-By: wanghao <sxmatch1986@gmail.com>

Closes-bug: #1809647

[1] https://review.openstack.org/#/c/620632/
[2] https://review.openstack.org/#/c/620632/2/cinder/api/common.py@356

Change-Id: I190cc67a001ffce2a533480e6e39e26cd3f981df