Fix retype notifier test case (BackupNFSTestCase)
When we do backup for a volume, the backup notification is called
periodically, but if we do other operations (like retype operation)
in parallel, there is a possible to record the "volume.retype" and
"backup.createprogress".

This patch try to mock the send notification methods to ensure this
doesn't happen.

Change-Id: Iad899944b54a0556ff0a946d18fcdb7e1680de69
Closes-Bug: #1803648