Refactor on policy in code
Recently, we add the policy in code test, and put all these
cases into test_volume_protection.py, but with the test case
growing up, it's difficult to maintenance.

In this patch, we refactor these test cases:
1. Add a new unit test case path - "tests/unit/policies/".
2. Split policy in code test in their own file, which is one-to-one
mapping to policy file.

Note that, there is no functional changes in this patch.

Change-Id: I0fe9167ace471e1dcc9c04823132d48665b5e927