Pass image_id to ImageDownloadFailed
When an ImageDownloadFailed exception is created in the
GlanceImageService, it's being passed the context whereas what
makes sense for the error message is the image_id.  This patch
makes the change and adds a unit test for a previous fix for
Bug #1798147 (which was the real point of this patch).

Change-Id: I6d8dedfd056add3414f8f4bf7f7279eae4763286
Closes-bug: #1811184
(cherry picked from commit 16b4346171d2027bb3c2630d5f28f6f002878870)