Refresh the Service.service_id after re-spawning children
Currently, the oslo.service provider a mechanism that the main process
will re-spawn the children as necessary. But if we kill the children
service, the services (except the latest service) are always down
after re-spawn.

This reason of this problem is that the Service.service_id is inherited
from the parent process [1] (the parent Service service id class
attribute) and has been recorded as the last created service [2][3]
when latest process is stared. But when re-spawning child process,
only start() method would be called[1], so that the Service.service_id
is not refreshed as expected.

In order to refresh the Service class attribute service_id, we
should store the service_id in instance attr origin_service_id, and set
the class attr back using the instance attribute in start method.

[1] https://github.com/openstack/oslo.service/blob/d987a4a/oslo_service/service.py#L648
[2] https://github.com/openstack/cinder/blob/099b141/cinder/service.py#L193
[3] https://github.com/openstack/cinder/blob/099b141/cinder/service.py#L344

Change-Id: Ibefda81215c5081634876a2064b15638388ae921
Closes-bug: #1811344
(cherry picked from commit 40127d95a9e83e97970dc388c18c99eea7715edc)