Use new target_* options in documentation
Various iscsi_* options were replaced with target_* options. The former
were initially deprecated and have now been removed.

This patch updates relevant documentation to use the new names.

There were some other hits in the documentation and code for these
names, but these appeared to be for third-party drivers, and I don't
have the context about which, if any, need to change.

Change-Id: Ic37ea85f4ec778644bc9d49f513e58dc28c87907