Create volume attachments status check.
Currently the new cinder attachment_create method allows volumes in
all status to be performed for single attachment. This may cause
problems if volume-attach operation is taken on volumes in creating
status. 
In order to avoid these problems, we should only allow volumes
in 'in-use', 'reserved' status to check whether this attachment
creation is a override operation for a same instance.

Related-bug: #1694530
Change-Id: Ie59f609d38290147a67a936c61e0164fe278da99
Closes-Bug: #1813267