PowerMax driver - handle special case where IG exists
An 'Initiator' does not show up in the initiatior list unless
it is part of a masking view or has previously been logged on.
However these initiators can still be part of an 'Initiator Group'
and an attempt to create a new 'Initiator Group' will fail as
an 'Initiator' cannot be belong to more than one 'Initiator Group'.
Handling this special case and bringing the initiator group check
before the storage group logic so rollback, if required, is tidier.

Change-Id: I62f8a5d75d82352875b179f3e6e1a3f343bb531d