Huawei driver refactor(2/10)
Huawei driver code is terribly decayed and hard to maintain.
From this patch on, we'll try to progressively refactor total
Huawei driver, make the code more clear, reliable and maintainable.

This patch mainly optimized the huawei_smartx.py and huawei_utils.py
files and the logic that according to it.

Change-Id: I132b917e338d43da1e7b46127b1f228cefb1c16a