Allow setting Pure host personality
Add a new configuration option for setting the Pure host personality.
It determines how the Purity system tunes the protocol used between
the array and the initiator. To ensure the array works optimally with
the host, set the personality to the name of the host operating or
virtual memory system. Valid values are aix, esxi, hitachi-vsp, hpux,
oracle-vm-server, solaris, and vms. If your system is not listed as
one of the valid host personalities, do not set the option. By
default, the host personality is not set.

Change-Id: I5065ccbed27c5b7c3b32ae8eda221c151a8bbaa9