Remove LOCI publishing from the post pipeline
Due to transitive rights changes in Zuul, the LOCI job that was
publishing Cinder containers after every build is now causing an error
when building the post pipeline graph that causes all post jobs to be
skipped. This removes the job for now until a better solution is found
by the team.

Change-Id: I77ed6e3caa5553193c3d886d86c52d23eb87b429
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>