Ignore hacking rule H101
I can't think of any examples when this has proven to
be truly useful.  Many of the entries point to
contributors who are no longer active in the community.

This rule also encourages people writing TODOs to put
their own name, which may needlessly prevent others
from helping fix issues in the code.

Change-Id: I6b749fcc459474eb6ca4ad6331de9138dd702cd0