VNX: update sg in cache
When `destroy_empty_storage_group` is enabled, the sg will be deleted
after the last lun is detached from it. The issue is that the sg cache
isn't updated, which caused the failure of subsequent attaching volumes
to the same host because the sg in the cache will be used directly which
doesn't exist any more.

The fix is updating the sg with poll after deleting it. This doesn't
delete the sg from cache explicitly but makes sure the sg in cache with
latest status `not-exist`. Subsequent attaching will create the sg if it
is `not-exist`.

Closes-bug: #1817385
Change-Id: Ibb39879a77c97c6a5d885461e93116810d16b265