RBD: save and restore multiattach features
This patch implements saving and restoring of RBD image features for
multiattach volume retyping operations.  If saved features are not found
because the volume was created with an earlier revision of this driver,
an exception is raised which causes the driver-optimized retype to
dropped in favor of volume migration.  In the migration, a new volume of
correct type is created and this volume will have the proper features
both disabled and stored as metadata.  No additional steps are required
to migrate from earlier code revisions.

Change-Id: If1e39293889212c77b025f5917ec9f41fd7d1384