Run cinderlib functional tests on Ceph job
This patch replaces job devstack-plugin-ceph-tempest with
cinder-plugin-ceph-tempest that inherits from it.

This job runs the same devstack-tempest.yaml playbook but also runs the
cinderlib-run.yaml playbook for the cinderlib functional tests.

Since the openstack/cinderlib project is a required project for this job
we'll be able to run cross-repo dependencies with cinderlib.

Change-Id: Ia532b3fcfbecee7c6dc7aa302e6d90a8067b7ff0