Raise eventlet lower-constraint to 0.22.0
Eventlet has a known issue when running under Python 3.6 [1]. This was
noticed in another OpenStack repo. As far as I know, we have not hit
this in Cinder. But since many drivers use requests and most are not
running third party CI under 3.6 yet, this is an attempt to avoid
running into an issue with the versions of eventlet that have this
issue.

[1] https://github.com/eventlet/eventlet/issues/371

Change-Id: Ia6139c052091eaee6b49a55f57385e650508a891
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>