Add upgrade checker for backup driver path
This adds an upgrade check to make sure the backup_driver setting is
using the full path to the driver and not just specifying the module.
Use of the module path for this setting was deprecated several releases
ago and removed in Stein.

Change-Id: I6b9722ab0d7a8bdf808bd937b0c237e3715ccaab
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>
(cherry picked from commit cef38b5e849bad294374adaf213efabfc26b0d5c)