Add upgrade check for presence of policy.json file
Correct file is policy.yaml and only needed if overriding defaults. This
only warns if the file is present in case it was left by a previous
version and not actually needed or used.

Also checks for the policy_file being overridden in config, and if so,
warns if the specified file does not exist.

Change-Id: Ia46e843ad245cf8263b97b773fac9bc6c6fe6647
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>
(cherry picked from commit 8c6355f3ccb64cb949a8280b6cd98a37e13bfaa7)