Remove non-exception from EXCEPTION_DETAIL_MAPPINGS
Removes a non-exception from the map and adds tests to check the
consistency of message_field.Action and .Detail fields.

Updates test-requirements and lower-constraints to use the most
recent version of ddt allowed by current upper-constraints.  (An
included test uses the idata decorator, which allows passing an
iterator to the test data.)

Closes-bug: #1822025
Change-Id: I0cad6589b145fac430942ccbb27279db7b3b152f