Remove scanning upgrade checker from opts generation
The upgrade checker is going to need to be able to load
configuration settings to check for things that will
interfere with an upgrade.  This will cause genopts
to incorrectly think an entry needs to be added in
cinder/opts.py .  This change add status.py to the
exception list.

Change-Id: Iab71497b5e9405429976b385dde2f6b7d585abae