Allow reset-state to detach volume as per policy
If the policy for os-reset_state is modified to RULE_ADMIN_OR_OWNER,
the owner(non-admin) is still not able to reset the volume
attach-status to detached because the db changes require admin context.
As seen in other cases (attachment_delete), the context is elevated
to allow it.
This patch adds the same support in-case of reset state.

Change-Id: Ib9c67a4893f34ef96caf1b808617a0a6b4cb8d5e
Closes-Bug: #1808113