Bump hacking version to 1.1.x
E731s are be fixed, since use of lambdas makes code
harder to debug.

Ignore E402 and W503/W504 since these don't make a
sensible case for us to change code.

This also requires fixing some E501 errors that should
have already been detected, but weren't.

E117 and E305 errors are disabled for the moment,
and are fixed in a subsequent patch.

Pin pycodestyle to 2.5.0.

Change-Id: Ia06940cff7c3b71938eddb271d2cb8170e02e954