Fix swift backup driver crashing during service startup
This patch fixes a crash in the swift backup driver during service
startup, when the backup_swift_url is None. When backup_swift_url is
None, the initialization code tries to fetch swift's endpoint from the
service catalog. However, this fails during service startup because the
context is missing the information required to access the catalog.

Closes-Bug: #1826434
Change-Id: I7e7dbdae72a08e663a96b6c01340ec2a5cbe883e