Make sure we install cinder requirements during the correct tox phase
When runnign tests under tox the virtualenv is created in a number of
phases (in brush strokes)
0. virtualenv -p $python .tox/$dir
1. pip install $deps
2. pip install .

Step 2 is run without any deps as they're all supposed to be installed
in phase 2

We have constraints in 'deps' we need to install all of cinders
requirements during phase 2.  If we don't then we get unconstrained
versions being installed during phase 2.

This shows up in a py27 run[1] as the absence of taskflow, and then in
[2] with new dependancies being installed

[1] http://logs.openstack.org/34/647934/3/check/openstack-tox-py27/93180f1/tox/py27-1.log
[2] http://logs.openstack.org/34/647934/3/check/openstack-tox-py27/93180f1/tox/py27-2.log

Change-Id: I5b1c285680f7c3256a707ea5973573f1176d4a26
(cherry picked from commit f12fda8ffbb42e71108701b38c09618997b836c3)