Add 512e/4k disk geometry configuration
Add support for configuring the physical_block_size and
logical_block_size hints of volumes using the new 'disk_geometry'
configuration option.

The infrastructure to configure this inside of nova/libvirt already
exists, however previously only the iSCSI/SolidFire driver set this
value. This provides a configurable option for all drivers.

In particular, Ceph performs much better when I/O is 4k-aligned
otherwise it has to read the 4k from disk, modify and write it again.
Linux guests generally submit 4k-aligned I/O, however Windows guests
generally submit 512b-aligned I/O. When hinted with
physical_block_size=4096 Windows guests will switch to submitting most
I/O as 4k-aligned.

The SolidFire driver stores the current configuration value in the
database against the volume's provider_geometry database field. That has
the effect of causing the configuration option to only apply to new
volumes and is then difficult to change the configuration after the
fact. Instead of storing it in the database, here we simply reference
the active configuration directly during volume connection so that a
configuration change is reflected next time a volume is used. Note that
rebooting an existing VM will not regenerate it's libvirt configuration
and you will need to do an operation which regenerates the libvirt XML
such as detaching/re-attaching the volume, cold migrating the VM or live
migrating (to update the configuration) and then rebooting (so that the
guests re-reads the information from the hardware) in order for the
change to take effect.

The default for new versions is disk_geometry=512, although it
recommended to use disk_geometry=512e which will become the default in
the 2027.1 release.

DocImpact
Change-Id: Ic746ebbac2c0f9ab66b5082d3c0bfa37830b3d94
Closes-Bug: 2127196
Signed-off-by: Zachary Raines <zachary.raines@canonical.com>