Revert "Implement volume capacity stats for VMware"
This reverts commit a742569dc96af4ceb5b6fb53300d8eafeffff306.

Commit a742569dc96af4ceb5b6fb53300d8eafeffff306 introduced
serious performance issues.

Our scale tests revealed ~20% degradation in various rally
tests after this commit. Our experiments show that the current
implementation does not scale well for a vCenter installation
with 30 datastores which is a smaller size deployment for
OpenStack clouds. The patch introduced a config option
'vmware_storage_profile'. When it is set, the driver makes
periodic expensive PBM API calls which put lot of load on
vCenter and the controller VM.

Change-Id: I838c32277624db3b82db65d7ecfe81e2a779e99a