Pure Storage FlashArray iSCSI driver target CIDR support
Add new configuration parameter to allow user to specify
the CIDR of iSCSI targets on a FlashArray that can be
connected to.
This allows a FlashArray to have targets in multiple VLANs
and Cinder not try to connect to inaccesible targets which
can cause serious timeouts when connecting Cinder volumes
to hosts.
Default setting will be 0.0.0.0/0 which allow connection
to any live iSCSI target.

Change-Id: Ic813445cb5d55edb68ca8e756b9350ca7e8a1270