LOG _init_vendor_properites
This patch adds an info log in _init_vendor_properties that will get
called when a driver doesn't directly implement the method.  It's
not currently required, but we want to highly encourage drivers
to implement this method and expose the capabilites the driver
supports, so admins can discover what extra specs are available.

Change-Id: Id4f7de01416cac48df31c6181befd0cd8c626631