Mark the Sheepdog driver unsupported
The Sheep dog test runs have not been passing
in the gate for quite sometime, to the point the
tests were moved to the experimental queue.

The most recent failures show that Sheepdog is not
supported on Bionic and mailing list posts show it
as no longer being maintained [1].  Given all
these datapoints it is time to more it
unsupported.

[1] http://lists.wpkg.org/pipermail/sheepdog/2019-March/068449.html

Change-Id: Id7bcacaaf41b3cc9657c78de26111c0996df8775