Mark the Nimble Driver Unsupported
The Nimble Driver 3rd Party CI hasn't run successfully
in 133 days.  This does not meet the requirements
for Cinder's 3rd Party CI requirements.

As a result the driver is being marked unsupported in the
Train release and should be removed in the 'U' release if
the problems are not resolved.

Change-Id: I801a887adf2db8cd5d14eaa77ebb1a1bf201ae99