Clean up test requirements
testresources and testscenarios were added in ec40c3b6 to address an
issue with oslo.db not pulling in its own requirements.

tempest was used when we had the tempest plugin in the main cinder repo
and was not removed when the plugin switched to its own repo.

os-api-ref is used for building the API reference and was a leftover
from before doc/requirements.txt was added.

Change-Id: Ib4fc7cb6199c9581c19d34ea43a2d15ec9003f13
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>