Update QoS and add Accuracy Note
We discussed at the last mid-cycle that the definition of QoS
should be updated to indicate this is for backend QoS support.
I have added that clarification with this patch.

I have also added a note that this is not an auto-generated
matrix and the accuracy should be handled accordingly.

Change-Id: Iccc3783997c2709dbc64bed8bd93d034e0fc2920