SF remove deprecated sf_allow_template_caching
The solidfire driver currently has built in template image caching. It
is not manageable through the normal OpenStack API's. SolidFires
template image caching is the main cause for the exception
DuplicateSFVolumeNames when the system is running in an HA
A/A configuration.

The current implementation has the configuration variable
sf_allow_template_caching, the module _create_image_volume, and
_verify_image_volume, which will be removed.

Cinder has the same functionality which is managed through the OpenStack
APIs and will not have the same HA A/A side effect of throwing
DuplicateSDuplicateSFVolumeNames exception.

Change-Id: I69ce8bda4302bf821739eb7276f2db12b7a259fe