Mark Datera driver unsupported
The Datera driver has not successfully run 3rd party CI for
a period greater than 60 days.  This falls out of the
Cinder requirements for 3rd Party CI.  As a result the
driver is being marked as unsupported in the Train release
and will be removed in the 'U' release if the problem is not
corrected.

Change-Id: If72c1430edfcb66e7d23c2c889b44c5459082414