Launch driver list generation with the same python interpreter
In package building process, there might be build environments
becoming more popular without Python 2.x being installed, with
other words /usr/bin/python does not exist. The existing script
uses a shebang of "env python", which works fine inside a virtualenv
but not outside. The easy solution is to just call the script
with the same python interpreter we're being called with.

Change-Id: Id91996545f7f75c1c62cea34e240b9655b8ff3af