Zadara VPSA: Fix driver force detach operation
Cinder API supports os-force_detach, which once invoked can call the
driver terminate_connection passing a 'None' connector. When receiving
this, the driver should handle this call by terminating every
connection between the specified volume and any host it is mapped to.
This patch addresses this behavior for the Zadara VPSA driver.

Volume detach 'force' flag is an internal flag due to an API change on
our side. It will not have any impact on the standard detach
procedure.

Change-Id: I09aeac3b777427f9694697f579c7fb9716e311e5