Compress images uploaded to Glance
When possible (based on the image format), compress
images that are uploaded to Glance.  For now, this is
done for the qcow2 format, but can be expanded to
other formats as needed.

This should be a performance win for some deployments,
consuming less network/storage at the expense of CPU.
An "image_compress_on_upload" option is provided to
disable this behavior for deployments that want to
prioritize minimal CPU usage, etc.

Closes-Bug: #1824821
Change-Id: I22bd2be6653bdc4e755ea8a9d8780e8e67b4485a