Never skip jobs when .zuul.yaml is changed
This is a partial revert of https://review.opendev.org/650138/

Currently it is possible to send a review which only touches
.zuul.yaml and changes the functional/tempest/grenade jobs
in a way that breaks them (for example, by removing 'voting: false',
or adding a new job of the said type, or altering their definition
with new parameters). Such changes would not be gated, thus
allowing it to be merged.

Change-Id: Id2b1476fbe0ae7bffd8f64e6f870bdc3721abbb8