Add "service token" documentation
Adds a doc about configuring and troubleshooting a service user
to send service tokens.

Also adds the config necessary for successful use of a service token
to the sample configuration file.  Only adds keystone v3 options
because v3 is required for service tokens.  Lists the service user
options before the session conf opts because the former are required
while the latter are less likely to be changed.

Closes-bug: #1837449

Change-Id: I4223624d00c713ab6b129ed278147f1c89995567