QNAP: Fix login on Python3
Method b64encode from base64 library returns different values for Python
2 and 3.

For Python 2 you get a string, but on Python 3 you get a bytes object,
which when converted to a string using the six.text_type method will
result in a string of the form "b'password'" instead of getting
"password".

This will result in posting a request with the wrong password and
failing to authenticate.

To fix this we only need to call the decode method on the result of
b64encode, as it works for both string and bytes objects and will always
return the string we want.

Closes-Bug: #1837538
Change-Id: I181fefbc8cc3d3fdfc996f4367658ea59a51a9fc