QNAP: Avoid unnecessary sleeps
QNAP driver is unnecessarily sleeping for a rety even when the operation
had succeeded and there is no need to sleep because we are not going to
retry.

This means that most operations will end up finishing 3 seconds later
than they should.

This patch makes it so that we don't do those unnecessary sleeps.

Change-Id: Iaafd72050a07d3f2736bea6d51c54d998d808aaa