Fix online data migrations
Online data migrations are part of the rolling upgrades mechanism, and
they are run on the current installed version before we upgrade to the
next one to ensure that all DB records have been upgraded even if they
have not been accessed during the life of the current release service
life.

This means that online migrations from one release should be removed on
the next release, as they are no longer needed.

We still have the Queens online data migrations in our code, and we
should remove them in master, Stein, and Rocky.

This patch also adds checks to the status tool to check that these
online data migrations have completed.

Closes-bug: #1837703
Change-Id: I025789f92eb318b8e5a531bea6cf106d52f268ae