NetApp SolidFire: Fix replication
This patch addresses several of the problems currently found in this
driver:

 - When creating a SolidFire volume with a volume type that has
replication enabled the replication_status of this volume was not
changing.
 - During initialization, if a cluster is already paired, the driver
fails to load.
 - Driver does not create volumes on the replication target on
several configurations. For example, if the volume is being
created from a snashot or from another volume, etc.
 - During Failback, the driver doesn't update iSCSI connection
info to the secondary cluster nor the volume's replication status.

Closes-bug: #1834013
Closes-bug: #1751932
Change-Id: I920f958db1d48e52b548082d852c03f427a279ca
Co-Authored-By: Fernando Ferraz <sfernand@netapp.com>
Co-Authored-By: Thiago Correa <thiago.correa@netapp.com>