NEC Driver: Storage assist retype and a bugfix
This patch adds storage assist retype support to NEC driver by
implementing retype(). In retype(), if the source and destination
backend are different, the driver calls a local migration function to
create a destination volume.
If they are the same, the driver updates QoS parameter without creating
a new volume. When creating a destination volume is completed and the
source volume is in NEC Storage, the driver marks the source volume as
deleting by adding a suffix '_d' to its LDN (Logical Disk Name in NEC
Storage). The driver used to delete the source volume on the
destination side, but it should be deleted on the source side later.
In the new implementation, the driver finds the deleting volume and
deletes it on the source side.

The patch also fixes the bug 1838955. The driver could not delete a
source volume when the destination volume was not in NEC Storage
because the driver used to delete the source volume on the destination
side. In the new implementation, The driver deletes the source volume
on the source side and the problem is solved.

Change-Id: I9e3b0dc9c9b6de393f73bd59ef26d28e75163b04
Closes-Bug: #1838955