Google backup: correct string encoding between py 2 and 3
Correcting encode() and decode() methods of 8-bit and Unicode
string objects.

In Python 2, Both 8-bit and Unicode strings had both methods
and the type of the return value was based on the specific
encoding passed in.

In Python 3, Text strings only have an encode() method, and
that method can only be used with codecs that produce bytes
objects. Similarly bytes and bytearray objects only have a
decode() method which can only be used with codecs that
produce string objects.

Change-Id: I214a786858655b8caf9612f3da5403c7d0bb7387
Closes-Bug: #1831735